backend: security authorize, call stored procedures

+++ core/src/main/java/com/gmail/mileshko/lesya/schedule/controller/AttendanceController.java
@@ -1,60 +1,68 @@
 package com.gmail.mileshko.lesya.schedule.controller;
 
-import com.gmail.mileshko.lesya.schedule.dto.AssessmentDto;
 import com.gmail.mileshko.lesya.schedule.dto.AttendanceDto;
 import com.gmail.mileshko.lesya.schedule.dto.NewAttendanceDto;
 import com.gmail.mileshko.lesya.schedule.entity.Student;
+import com.gmail.mileshko.lesya.schedule.entity.User;
 import com.gmail.mileshko.lesya.schedule.exception.NoSuchEntityException;
-import com.gmail.mileshko.lesya.schedule.repository.StudentRepository;
 import com.gmail.mileshko.lesya.schedule.service.AttendanceService;
 import com.gmail.mileshko.lesya.schedule.service.StudentService;
+import com.gmail.mileshko.lesya.schedule.service.UserService;
 import com.gmail.mileshko.lesya.schedule.util.Mapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
+import static com.gmail.mileshko.lesya.schedule.security.SecurityConstants.HEADER_STRING;
+
 @RestController
 @RequestMapping("attendance")
 public class AttendanceController {
-    private  final StudentService studentService;
+    private final StudentService studentService;
     private final AttendanceService attendanceService;
-    private final StudentRepository studentRepository;
+    private final UserService userService;
 
 
     @Autowired
-    public AttendanceController(StudentService studentService, AttendanceService attendanceService, StudentRepository studentRepository) {
+    public AttendanceController(StudentService studentService, AttendanceService attendanceService, UserService userService) {
         this.studentService = studentService;
         this.attendanceService = attendanceService;
-        this.studentRepository = studentRepository;
+        this.userService = userService;
     }
 
 
     @GetMapping("get")
-    public List<AttendanceDto> getAttendance(@RequestHeader("token") String token) throws NoSuchEntityException {
-        Student student = studentService.validate(token);
+    @PreAuthorize("hasRole('STUDENT')")
+    public List<AttendanceDto> getAttendance(@RequestHeader(HEADER_STRING) String token) throws NoSuchEntityException {
+        User user = userService.getUser(token);
+        Student student = studentService.getStudent(user);
         return Mapper.mapAll(attendanceService.getAttendance(student), AttendanceDto.class);
-
     }
 
     @GetMapping("student")
+    @PreAuthorize("hasRole('STUDENT')")
     public List<AttendanceDto> getStudentAttendance(@RequestParam("studentId") Long id) throws NoSuchEntityException {
-        Student student = studentRepository.findById(id).orElseThrow(()-> new NoSuchEntityException("Студент не найден") );
+        Student student = studentService.getStudentById(id);
         return Mapper.mapAll(attendanceService.getAttendance(student), AttendanceDto.class);
     }
 
 
     @PostMapping("add")
-    public void addAttendance(@RequestBody NewAttendanceDto newAttendanceDto, @RequestHeader("token") String token) throws NoSuchEntityException {
-        Student headman = studentService.validate(token);
-        studentService.isHeadman(headman);
+    public void addAttendance(@RequestBody NewAttendanceDto newAttendanceDto, @RequestHeader(HEADER_STRING) String token) throws Throwable {
+        User user = userService.getUser(token);
+        Student student = studentService.getStudent(user);
+        studentService.isHeadman(student);
         attendanceService.addAttendance(newAttendanceDto);
     }
 
     @PostMapping("delete")
-    public void deleteAttendance(@RequestBody Long attendanceId, @RequestHeader("token") String token) throws NoSuchEntityException {
-        Student headman = studentService.validate(token);
-        studentService.isHeadman(headman);
-        attendanceService.deleteAttendance(attendanceId);
+    public void deleteAttendance(@RequestBody Long attendanceId, @RequestHeader(HEADER_STRING) String token) throws NoSuchEntityException {
+        User user = userService.getUser(token);
+        Student student = studentService.getStudent(user);
+        if (studentService.isHeadman(student)) {
+            attendanceService.deleteAttendance(attendanceId);
+        }
     }
 }
